Estimating multipliers to fit growth curves
A first attempt was made using the multiplier by the SC (species
composition). It worked for all species except Pl. A SC muliplier
function would not work because there are BA values that cannot be
achieved with a given SC even with a multiplier. I then multiplied the
entire BAinc function by a factor in the Pl case. I attempted the same
approach for the other species, but it did not work. I then decided to
use the multiplier factor times SCs in Aw, SB, and Sw, and used the
multiplier times the entire BAinc function for Pl.  In other words, I
had to alter the shape of the growth curve and used the combination
that worked.