Tested with different bhages
General conclusion:
there is an increasing difference in results when estimating BA using
different methods after time t for younger plot. For older plots,
differences reduce considerably.

There is a problem when plots are very young (bhage ˜10 years). Very
little can be said other than their current status and predictions
become uncertain.

Compounded with this, there is an algorithm problem that sets the
densities to zero that needs to be solved.