Aspen converging - a few problems yet
Behaviour of curves after time T seems to be sometimes closer to gypsy
using the “no” routine and sometime “yes”routine. I need to understand
when is it better using no and using yes and apply the rule. One
alternative is using no only for Aspen. That might do the trick.