Add definitive profiling for simulate_forwards_df
This glosses over the higher level things that happen, like module
loading, i/o, etc. But, it was required to isolate the performance of
the simulation from the other things - previous analyses were too
difficult to parse.

The next step is to use this information to improve performance of
simulation.