Factor density function repetition out of sdf
The sdf functions require variables local to the density functions.
In the future, we can create functions that generate density function
with fixed parameters except for sdf, and use scipy to solve those
functions for sdf. But that process is a bit more time consuming and
prone to changes in the reference test data. This at least keeps things
more DRY and helps with maintainability and bug reduction.