Do not distribute data with built package
Some of the test were failing on codeship and in fresh images because
the inclusion of package data was not correct. package data paths have
to be relative to packages, i.e. dirs with __init__.py included in
setup.py as packages.

The current python packaging conventions is to NOT distributes tests and
test data with the package - that's for users. Only devs need those
data.

So, revert to using -e install in Dockerfile and add build as a step in
codeship.