Fix Sw simulation function
for #66. To chacterize this: the simulation goes as follows

until the time of data, the quantity of species composition * increases
exponentially

from the time of data to the end of the simulation, the estimate species
composition, from the density simulation, is used

the correction factor may or may not be used in these future years (e.g.
for aw it is used)

note that it should be optional for the species comp and correction
factor to increase exponentially for the whole simulation

The naive thought is that the estimated densities and species
compositions from 0 to the time of the data could be used

Generate test array
Remove debug