Make basal area sim compat with skip backwards sim
This is a hack but should work well for time being

Basal area simulation required some logic to determine whether it should
use the bhage/data year parameters for calculating increment.
It also used the year of the simulation, well, the counter of the years.
This is not compatible with skipping the backwards simulation.

One way to control it would have been to keep the year, instead of the
counter. Pretty laborious.

The source of the issue is that the parameters for basal area increment
should be determined elsewhere. Probably in their generation in the
densities_speciescomp_topheight function or before or after that
function. These basal area simulation functions should just use those
parameters, and not be responsible for overriding them.